requirements.txt
aiohttp==3.8.1
aioice==0.7.6
aiortc==1.3.0
aiosignal==1.2.0
altgraph==0.17.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
av==8.1.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cx-Freeze==6.10
cx-Logging==3.0
cycler==0.11.0
decorator==4.4.2
dlib==19.23.1
dnspython==2.2.0
eventlet==0.33.0
ffmpeg==1.4
Flask==2.0.3
Flask-JSGlue==0.3.1
Flask-SocketIO==4.3.2
fonttools==4.29.1
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
google-crc32c==1.3.0
greenlet==1.1.2
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.1
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lief==0.11.5
MarkupSafe==2.0.1
matplotlib==3.5.1
moviepy==1.0.3
multidict==6.0.2
netifaces==0.11.0
networkx==2.6.3
numpy==1.21.5
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pefile==2021.9.3
Pillow==8.4.0
proglog==0.1.9
pycparser==2.21
pydub==0.25.1
pyee==9.0.4
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
pylibsrtp==0.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
PyWavelets==1.2.0
pywin32-ctypes==0.2.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
threaded==4.1.0
tifffile==2021.11.2
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
